+++ src/support.js
@@ -5,23 +5,38 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
     return Cypress.browser && Cypress.browser.isHeaded;
   }
 
-  function shouldFailFast() {
-    return (
-      Cypress.env(PLUGIN_ENVIRONMENT_VAR) === true ||
-      Cypress.env(PLUGIN_ENVIRONMENT_VAR) === "true"
-    );
+  function getFailFastEnvironmentConfig() {
+    return {
+      enabled:
+        Cypress.env(PLUGIN_ENVIRONMENT_VAR) === true ||
+        Cypress.env(PLUGIN_ENVIRONMENT_VAR) === "true",
+    };
   }
 
-  function testState(test) {
-    return (
-      test.currentTest &&
-      test.currentTest.state === "failed" &&
-      test.currentTest.currentRetry() === test.currentTest.retries()
-    );
+  function getTestFailFastConfig(currentTest) {
+    if (currentTest.cfg && currentTest.cfg.failFast) {
+      return currentTest.cfg.failFast;
+    }
+    if (currentTest.parent) {
+      return getTestFailFastConfig(currentTest.parent);
+    }
+    return getFailFastEnvironmentConfig();
+  }
+
+  function pluginIsEnabled() {
+    return getFailFastEnvironmentConfig().enabled;
+  }
+
+  function shouldSkipRestOfTests(currentTest) {
+    return getTestFailFastConfig(currentTest).enabled;
+  }
+
+  function testHasFailed(currentTest) {
+    return currentTest.state === "failed" && currentTest.currentRetry() === currentTest.retries();
   }
 
   beforeEach(function () {
-    if (shouldFailFast()) {
+    if (pluginIsEnabled()) {
       cy.task("shouldSkipDueToFailFast").then((value) => {
         if (value === true) {
           Cypress.runner.stop();
@@ -32,14 +47,15 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
 
   afterEach(function () {
     // Mark skip flag as true if test failed
-    if (shouldFailFast() && testState(this)) {
+    const currentTest = this.currentTest;
+    if (currentTest && testHasFailed(currentTest) && shouldSkipRestOfTests(currentTest)) {
       cy.task("shouldSkipDueToFailFast", true);
       Cypress.runner.stop();
     }
   });
 
   before(function () {
-    if (isHeaded() && shouldFailFast()) {
+    if (isHeaded() && pluginIsEnabled()) {
       // Reset the shouldSkip flag at the start of a run, so that it
       //  doesn't carry over into subsequent runs.
       // Do this only for headed runs because in headless runs,